moved client to its own module
